Reworking of the Robot init to minimize statics.
Static variables make it easy bind and reference variables
as with Commands and Subsytems, but makes it makes the code
very brittle.  This led us to the place where it's unclear
which objects are in which state during the init sequence.

This change reworks the initialization as follows:

* Robot is converted to a singleton, obtained via Robot.Get() or
  Robot.getInstance()

* Robot first contructs OI, since the buttons, and other channel
  IDs are of value to subsystems.

* Robot next constructs the subsystems, passing references to
  its member variables. Subsytems should *not* reach back into
  the Robot instance to pull on values.

* Initialization sequences are now proected by try/catch blocks.
  This approach is theoretically more robust that the ModuleManager
  approach recently introduced.  I've spoken with Binnur about this
  and she's tenatively on the same page: the try/catch approach
  seems preferrable. It may be that we don't have enough blocks,
  as usual coding is done without access to a real robot.

There are still lots of issues with the code organization pertaining
to ecapsulation. Basically we'd like to grant access to things on
a fairly strict "need-to-know" policy.  This makes it much easier to
add new systems and change existing behaviors, etc.